refactor: add shared test for delete many with prefix
