+++ tests/test_redis_cache.py
@@ -1,6 +1,7 @@
 import pytest
 from clear import ClearTests
 from common import CommonTests
+from delete_many_with_prefix import DeleteManyWithPrefixTests
 from has import HasTests
 
 from cachelib import RedisCache
@@ -43,17 +44,10 @@ def my_callable_key() -> str:
 
 
 @pytest.mark.usefixtures("redis_server")
-class TestRedisCache(CommonTests, ClearTests, HasTests):
+class TestRedisCache(CommonTests, ClearTests, HasTests, DeleteManyWithPrefixTests):
     def test_callable_key(self):
         cache = self.cache_factory()
         assert cache.set(my_callable_key, "sausages")
         assert cache.get(my_callable_key) == "sausages"
         assert cache.set(lambda: "spam", "sausages")
         assert cache.get(lambda: "spam") == "sausages"
-
-    def test_delete_many_with_prefix(self):
-        cache = self.cache_factory(key_prefix="test_prefix:")
-        cache.set_many({"foo": 1, "bar": 2, "baz": 3})
-        result = cache.delete_many("foo", "bar", "baz")
-        assert result == ["foo", "bar", "baz"]
-        assert not any(cache.get_many("foo", "bar", "baz"))
